arrayのテンプレートパラメータをsize_tに変更、iic->i2cの名称変更、format_stringのデフォルトパラメータ追加

+++ TankController/array.hpp
@@ -4,7 +4,7 @@
 namespace util
 {
 	
-template<class T, int N>
+template<class T, size_t N>
 class array
 {
 	private:
